+++ src/MagellanServiceProvider.php
@@ -3,10 +3,8 @@
 namespace Clickbar\Magellan;
 
 use Clickbar\Magellan\Commands\UpdatePostgisColumns;
-use Clickbar\Magellan\Data\Geometries\Geometry;
 use Clickbar\Magellan\Data\Geometries\GeometryFactory;
 use Clickbar\Magellan\Database\Builder\BuilderMacros;
-use Clickbar\Magellan\IO\Generator\WKT\WKTGenerator;
 use Clickbar\Magellan\IO\GeometryModelFactory;
 use Clickbar\Magellan\IO\Parser\Geojson\GeojsonParser;
 use Clickbar\Magellan\IO\Parser\WKB\WKBParser;
@@ -17,7 +15,6 @@ use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Query\Builder;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Schema\Grammars\PostgresGrammar;
-use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\DB;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
@@ -66,27 +63,6 @@ class MagellanServiceProvider extends PackageServiceProvider
             DB::registerDoctrineType(\Clickbar\Magellan\DBAL\Types\GeometryType::class, 'geometry', 'geometry');
             DB::registerDoctrineType(\Clickbar\Magellan\DBAL\Types\GeographyType::class, 'geography', 'geography');
         }
-
-        /**
-         * We use the beforeExecuting event to convert Geometry objects that are passed as bindings
-         * to the sql string representation using the configured generator.
-         *
-         * This will allow us to explicitly use another generator than the string generator, because
-         * Laravel would otherwise convert the Geometry object to a string using the __toString() method.
-         */
-        DB::beforeExecuting(function ($query, &$bindings, $connection) {
-            $postgisSchema = Config::get('magellan.schema', 'public');
-            $generatorClass = Config::get('magellan.sql_generator', WKTGenerator::class);
-
-            /** @var \Clickbar\Magellan\IO\Generator\BaseGenerator */
-            $generator = new $generatorClass();
-
-            foreach ($bindings as $key => $value) {
-                if ($value instanceof Geometry) {
-                    $bindings[$key] = $generator->toPostgisGeometrySql($value, $postgisSchema);
-                }
-            }
-        });
     }
 
     private function registerBuilderMixin($mixin)